requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.19
langchain-core==0.3.63
langchain-chroma==0.2.4
openai==1.82.0
flashrank==0.2.10
streamlit==1.45.1
schedule==1.2.2
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.13.4
pydantic==2.11.4
chromadb==1.0.10
selenium==4.32.0
crwl==0.0.2
python-dotenv==1.0.1
mcp[cli]>=1.0.0
